+++ src/generate/mod.rs
@@ -1,537 +1,17 @@
-use std::collections::HashMap;
 use std::io::{self, Write};
 
 use cast::u64;
 use quote::{ToTokens, Tokens};
-use svd::{Access, BitRange, Defaults, Device, EnumeratedValues, Field, Peripheral, Register,
+use svd::{Access, BitRange, Defaults, EnumeratedValues, Field, Peripheral, Register,
           Usage, WriteConstraint};
 use syn::{self, Ident};
 
 use errors::*;
 use util::{self, ToSanitizedSnakeCase, ToSanitizedUpperCase, U32Ext, BITS_PER_BYTE};
-use Target;
 
-/// Whole device generation
-pub fn device(d: &Device, target: &Target, items: &mut Vec<Tokens>) -> Result<()> {
-    let doc = format!(
-        "Peripheral access API for {0} microcontrollers \
-         (generated using svd2rust v{1})\n\n\
-         You can find an overview of the API [here].\n\n\
-         [here]: https://docs.rs/svd2rust/{1}/svd2rust/#peripheral-api",
-        d.name.to_uppercase(),
-        env!("CARGO_PKG_VERSION")
-    );
-
-    if *target == Target::Msp430 {
-        items.push(quote! {
-            #![feature(abi_msp430_interrupt)]
-        });
-    }
-
-    if *target != Target::None {
-        items.push(quote! {
-            #![cfg_attr(feature = "rt", feature(global_asm))]
-            #![cfg_attr(feature = "rt", feature(macro_reexport))]
-            #![cfg_attr(feature = "rt", feature(used))]
-        });
-    }
-
-    items.push(quote! {
-        #![doc = #doc]
-        #![allow(private_no_mangle_statics)]
-        #![deny(missing_docs)]
-        #![deny(warnings)]
-        #![allow(non_camel_case_types)]
-        #![feature(const_fn)]
-        #![no_std]
-    });
-
-    match *target {
-        Target::CortexM => {
-            items.push(quote! {
-                extern crate cortex_m;
-                #[macro_reexport(default_handler, exception)]
-                #[cfg(feature = "rt")]
-                extern crate cortex_m_rt;
-            });
-        }
-        Target::Msp430 => {
-            items.push(quote! {
-                extern crate msp430;
-                #[macro_reexport(default_handler)]
-                #[cfg(feature = "rt")]
-                extern crate msp430_rt;
-            });
-        }
-        Target::None => {}
-    }
-
-    items.push(quote! {
-        extern crate bare_metal;
-        extern crate vcell;
-
-        use core::ops::Deref;
-        use core::marker::PhantomData;
-    });
-
-    if let Some(cpu) = d.cpu.as_ref() {
-        let bits = util::unsuffixed(cpu.nvic_priority_bits as u64);
-
-        items.push(quote! {
-            /// Number available in the NVIC for configuring priority
-            pub const NVIC_PRIO_BITS: u8 = #bits;
-        });
-    }
-
-    ::generate::interrupt(d, target, &d.peripherals, items);
-
-    const CORE_PERIPHERALS: &[&str] = &[
-        "CBP",
-        "CPUID",
-        "DCB",
-        "DWT",
-        "FPB",
-        "FPU",
-        "ITM",
-        "MPU",
-        "NVIC",
-        "SCB",
-        "SYST",
-        "TPIU",
-    ];
-
-    let mut fields = vec![];
-    let mut exprs = vec![];
-    if *target == Target::CortexM {
-        items.push(quote! {
-            pub use cortex_m::peripheral::Peripherals as CorePeripherals;
-        });
-
-        // NOTE re-export only core peripherals available on *all* Cortex-M devices
-        // (if we want to re-export all core peripherals available for the target then we are going
-        // to need to replicate the `#[cfg]` stuff that cortex-m uses and that would require all
-        // device crates to define the custom `#[cfg]`s that cortex-m uses in their build.rs ...)
-        items.push(quote! {
-            pub use cortex_m::peripheral::CPUID;
-            pub use cortex_m::peripheral::DCB;
-            pub use cortex_m::peripheral::DWT;
-            pub use cortex_m::peripheral::MPU;
-            pub use cortex_m::peripheral::NVIC;
-            pub use cortex_m::peripheral::SCB;
-            pub use cortex_m::peripheral::SYST;
-        });
-    }
-
-    for p in &d.peripherals {
-        if *target == Target::CortexM && CORE_PERIPHERALS.contains(&&*p.name.to_uppercase()) {
-            // Core peripherals are handled above
-            continue;
-        }
-
-        ::generate::peripheral(p, &d.peripherals, items, &d.defaults)?;
-
-        if p.registers
-            .as_ref()
-            .map(|v| &v[..])
-            .unwrap_or(&[])
-            .is_empty() && p.derived_from.is_none()
-        {
-            // No register block will be generated so don't put this peripheral
-            // in the `Peripherals` struct
-            continue;
-        }
-
-        let p = p.name.to_sanitized_upper_case();
-        let id = Ident::new(&*p);
-        fields.push(quote! {
-            #[doc = #p]
-            pub #id: #id
-        });
-        exprs.push(quote!(#id: #id { _marker: PhantomData }));
-    }
-
-    let take = match *target {
-        Target::CortexM => Some(Ident::new("cortex_m")),
-        Target::Msp430 => Some(Ident::new("msp430")),
-        Target::None => None,
-    }.map(|krate| quote! {
-        /// Returns all the peripherals *once*
-        #[inline]
-        pub fn take() -> Option<Self> {
-            #krate::interrupt::free(|_| {
-                if unsafe { DEVICE_PERIPHERALS } {
-                    None
-                } else {
-                    Some(unsafe { Peripherals::steal() })
-                }
-            })
-        }
-    });
-
-    items.push(quote! {
-        // NOTE `no_mangle` is used here to prevent linking different minor versions of the device
-        // crate as that would let you `take` the device peripherals more than once (one per minor
-        // version)
-        #[no_mangle]
-        static mut DEVICE_PERIPHERALS: bool = false;
-
-        /// All the peripherals
-        #[allow(non_snake_case)]
-        pub struct Peripherals {
-            #(#fields,)*
-        }
-
-        impl Peripherals {
-            #take
-
-            /// Unchecked version of `Peripherals::take`
-            pub unsafe fn steal() -> Self {
-                debug_assert!(!DEVICE_PERIPHERALS);
-
-                DEVICE_PERIPHERALS = true;
-
-                Peripherals {
-                    #(#exprs,)*
-                }
-            }
-        }
-    });
-
-    Ok(())
-}
-
-/// Generates code for `src/interrupt.rs`
-pub fn interrupt(
-    device: &Device,
-    target: &Target,
-    peripherals: &[Peripheral],
-    items: &mut Vec<Tokens>,
-) {
-    let interrupts = peripherals
-        .iter()
-        .flat_map(|p| p.interrupt.iter())
-        .map(|i| (i.value, i))
-        .collect::<HashMap<_, _>>();
-
-    let mut interrupts = interrupts.into_iter().map(|(_, v)| v).collect::<Vec<_>>();
-    interrupts.sort_by_key(|i| i.value);
-
-    let mut arms = vec![];
-    let mut elements = vec![];
-    let mut names = vec![];
-    let mut variants = vec![];
-
-    // Current position in the vector table
-    let mut pos = 0;
-    let mut mod_items = vec![];
-    mod_items.push(quote! {
-        use bare_metal::Nr;
-    });
-    for interrupt in &interrupts {
-        while pos < interrupt.value {
-            elements.push(quote!(None));
-            pos += 1;
-        }
-        pos += 1;
-
-        let name_uc = Ident::new(interrupt.name.to_sanitized_upper_case());
-        let description = format!(
-            "{} - {}",
-            interrupt.value,
-            interrupt
-                .description
-                .as_ref()
-                .map(|s| util::respace(s))
-                .unwrap_or_else(|| interrupt.name.clone())
-        );
-
-        let value = util::unsuffixed(u64(interrupt.value));
-
-        variants.push(quote! {
-            #[doc = #description]
-            #name_uc,
-        });
-
-        arms.push(quote! {
-            Interrupt::#name_uc => #value,
-        });
-
-        elements.push(quote!(Some(#name_uc)));
-        names.push(name_uc);
-    }
-
-    let aliases = names
-        .iter()
-        .map(|n| {
-            format!(
-                "
-.weak {0}
-{0} = DH_TRAMPOLINE",
-                n
-            )
-        })
-        .collect::<Vec<_>>()
-        .concat();
-
-    let n = util::unsuffixed(u64(pos));
-    match *target {
-        Target::CortexM => {
-            let is_armv6 = match device.cpu {
-                Some(ref cpu) => cpu.name.starts_with("CM0"),
-                None => true, // default to armv6 when the <cpu> section is missing
-            };
-
-            if is_armv6 {
-                // Cortex-M0(+) are ARMv6 and don't have `b.w` (branch with 16 MB range). This
-                // can cause linker errors when the handler is too far away. Instead of a small
-                // inline assembly shim, we generate a function for those targets and let the
-                // compiler do the work (sacrificing a few bytes of code).
-                mod_items.push(quote! {
-                    #[cfg(feature = "rt")]
-                    extern "C" {
-                        fn DEFAULT_HANDLER();
-                    }
-
-                    #[cfg(feature = "rt")]
-                    #[allow(non_snake_case)]
-                    #[no_mangle]
-                    pub unsafe extern "C" fn DH_TRAMPOLINE() {
-                        DEFAULT_HANDLER();
-                    }
-                });
-            } else {
-                mod_items.push(quote! {
-                    #[cfg(all(target_arch = "arm", feature = "rt"))]
-                    global_asm!("
-                    .thumb_func
-                    DH_TRAMPOLINE:
-                        b DEFAULT_HANDLER
-                    ");
-
-                    /// Hack to compile on x86
-                    #[cfg(all(target_arch = "x86_64", feature = "rt"))]
-                    global_asm!("
-                    DH_TRAMPOLINE:
-                        jmp DEFAULT_HANDLER
-                    ");
-                })
-            }
-
-            mod_items.push(quote! {
-                #[cfg(feature = "rt")]
-                global_asm!(#aliases);
-
-                #[cfg(feature = "rt")]
-                extern "C" {
-                    #(fn #names();)*
-                }
-
-                #[allow(private_no_mangle_statics)]
-                #[cfg(feature = "rt")]
-                #[doc(hidden)]
-                #[link_section = ".vector_table.interrupts"]
-                #[no_mangle]
-                #[used]
-                pub static INTERRUPTS: [Option<unsafe extern "C" fn()>; #n] = [
-                    #(#elements,)*
-                ];
-            });
-        }
-        Target::Msp430 => {
-            mod_items.push(quote! {
-                #[cfg(feature = "rt")]
-                global_asm!("
-                DH_TRAMPOLINE:
-                    jmp DEFAULT_HANDLER
-                ");
-
-                #[cfg(feature = "rt")]
-                global_asm!(#aliases);
-
-                #[cfg(feature = "rt")]
-                extern "msp430-interrupt" {
-                    #(fn #names();)*
-                }
-
-                #[allow(private_no_mangle_statics)]
-                #[cfg(feature = "rt")]
-                #[doc(hidden)]
-                #[link_section = ".vector_table.interrupts"]
-                #[no_mangle]
-                #[used]
-                pub static INTERRUPTS:
-                    [Option<unsafe extern "msp430-interrupt" fn()>; #n] = [
-                        #(#elements,)*
-                    ];
-            });
-        }
-        Target::None => {}
-    }
-
-    mod_items.push(quote! {
-        /// Enumeration of all the interrupts
-        pub enum Interrupt {
-            #(#variants)*
-        }
-
-        unsafe impl Nr for Interrupt {
-            #[inline]
-            fn nr(&self) -> u8 {
-                match *self {
-                    #(#arms)*
-                }
-            }
-        }
-    });
-
-    if *target != Target::None {
-        let abi = match *target {
-            Target::Msp430 => "msp430-interrupt",
-            _ => "C",
-        };
-        mod_items.push(quote! {
-            #[cfg(feature = "rt")]
-            #[macro_export]
-            macro_rules! interrupt {
-                ($NAME:ident, $path:path, locals: {
-                    $($lvar:ident:$lty:ty = $lval:expr;)*
-                }) => {
-                    #[allow(non_snake_case)]
-                    mod $NAME {
-                        pub struct Locals {
-                            $(
-                                pub $lvar: $lty,
-                            )*
-                        }
-                    }
-
-                    #[allow(non_snake_case)]
-                    #[no_mangle]
-                    pub extern #abi fn $NAME() {
-                        // check that the handler exists
-                        let _ = $crate::interrupt::Interrupt::$NAME;
-
-                        static mut LOCALS: self::$NAME::Locals =
-                            self::$NAME::Locals {
-                                $(
-                                    $lvar: $lval,
-                                )*
-                            };
-
-                        // type checking
-                        let f: fn(&mut self::$NAME::Locals) = $path;
-                        f(unsafe { &mut LOCALS });
-                    }
-                };
-                ($NAME:ident, $path:path) => {
-                    #[allow(non_snake_case)]
-                    #[no_mangle]
-                    pub extern #abi fn $NAME() {
-                        // check that the handler exists
-                        let _ = $crate::interrupt::Interrupt::$NAME;
-
-                        // type checking
-                        let f: fn() = $path;
-                        f();
-                    }
-                }
-            }
-        });
-    }
-
-    if interrupts.len() > 0 {
-        items.push(quote! {
-            pub use interrupt::Interrupt;
-
-            #[doc(hidden)]
-            pub mod interrupt {
-                #(#mod_items)*
-            }
-        });
-    }
-}
-
-pub fn peripheral(
-    p: &Peripheral,
-    all_peripherals: &[Peripheral],
-    items: &mut Vec<Tokens>,
-    defaults: &Defaults,
-) -> Result<()> {
-    let name_pc = Ident::new(&*p.name.to_sanitized_upper_case());
-    let address = util::hex(p.base_address);
-    let description = util::respace(p.description.as_ref().unwrap_or(&p.name));
-
-    let name_sc = Ident::new(&*p.name.to_sanitized_snake_case());
-    let (base, derived) = if let Some(base) = p.derived_from.as_ref() {
-        // TODO Verify that base exists
-        // TODO We don't handle inheritance style `derivedFrom`, we should raise
-        // an error in that case
-        (Ident::new(&*base.to_sanitized_snake_case()), true)
-    } else {
-        (name_sc.clone(), false)
-    };
-
-    items.push(quote! {
-        #[doc = #description]
-        pub struct #name_pc { _marker: PhantomData<*const ()> }
-
-        unsafe impl Send for #name_pc {}
-
-        impl #name_pc {
-            /// Returns a pointer to the register block
-            pub fn ptr() -> *const #base::RegisterBlock {
-                #address as *const _
-            }
-        }
-
-        impl Deref for #name_pc {
-            type Target = #base::RegisterBlock;
-
-            fn deref(&self) -> &#base::RegisterBlock {
-                unsafe { &*#name_pc::ptr() }
-            }
-        }
-    });
-
-    if derived {
-        return Ok(())
-    }
-
-    let registers = p.registers.as_ref().map(|x| x.as_ref()).unwrap_or(&[][..]);
-
-    // No `struct RegisterBlock` can be generated
-    if registers.is_empty() {
-        // Drop the `pub const` definition of the peripheral
-        items.pop();
-        return Ok(());
-    }
-
-    let mut mod_items = vec![];
-    mod_items.push(::generate::register_block(registers, defaults)?);
-
-    for register in registers {
-        ::generate::register(
-            register,
-            registers,
-            p,
-            all_peripherals,
-            defaults,
-            &mut mod_items,
-        )?;
-    }
-
-    let description = util::respace(p.description.as_ref().unwrap_or(&p.name));
-    items.push(quote! {
-        #[doc = #description]
-        pub mod #name_sc {
-            use vcell::VolatileCell;
-
-            #(#mod_items)*
-        }
-    });
-
-    Ok(())
-}
+pub mod device;
+pub mod interrupt;
+pub mod peripheral;
 
 struct RegisterBlockField {
     field: syn::Field,
